[iOS] refactor: CellRegistration 부분을 담당하는 객체 생성
